Add TADO_MODE overlay constant
Previously the CONST_OVERLAY_TADO_MODE was set to "NEXT_TIME_BLOK". However, there is also an overlay type called "TADO_MODE".
The CONST_OVERLAY_TADO_MODE is now set to "TADO_MODE" and a new CONST_OVERLAY_NEXT_TIME_BLOCK is introduced. Furthermore, an elaborative comment on the overlay modes/types was added.
Since this is a breaking change for everyone using the previous definition of CONST_OVERLAY_TADO_MODE, it should be discussed how to proceed. However, using a const named this way might have caused unintended behavior in the past.